Replace OSH-AIO instructions and add links
Currently the docs here have the instructions from the Openstack-Helm
All-in-One environment docs instead of just linking to the docs
themselves. This change removes the redundant instructions and replaces
them with a link to the source of the instructions. This also adds a
link to the Airship in a Bottle instructions to the postgresql-utility
docs.

This is beneficial both to reduce the documentation here, as well as
avoiding changes to the original documents not making their way here.

Change-Id: I54af601112f93d9b50653afb56b93d0cd253658e