Add the capability to create a test database
In this patchset, if configured to do so, a test database with
a table and a couple of rows in the table is created at bootstrap time
of the mysqlclient utility container.
The test database will be created for each of the Mariadb DBs in the
mariadb backup restore enabled namespaces.

Change-Id: I3afb85debfafde0585abd726fe8b2f999f590294