Allow dbutils user to specify node in each command
dbutils can still use NODE as an exported env however, this
modification allows the user to overwrite the NODE with each
command if the user would like to use a different NODE between
commands or in the case that NODE is not exported prior to running
the script.

Also set ONDEMAND_JOB to match how we also set the env in
postgres and mariadb rather than expect it to already be exported.

Change-Id: I7afc346d34bcc6b03489455189138cf1cd1a61d4