Fix backup/restore bug for remote backup disabled case
Currently, if you try to start an on-demand container using the
"utilscli dbutils" command, on a site where remote backup has been
disabled, then the on-demand container will get a container create error
saying that it cannot read the backup-user secret (which does not get
generated when remote backup is disabled).

This patchset disables the reading of variables from the backup-user
secret (aka, the rgw secret) if remote backup is disabled.

Change-Id: I326a5b812f6b13d7dd42f4d3e339ba0d37eef538