Persistent On-Demand flag and db locking mechanism on backup and restore.
Currently when not using interactive mode, the dbutils will remove the
On-Demand pod after each command. This is not efficient if there is a need to
run multiple command in succession. By using the -p flag, the On-Demand pod
will not be deleted after the task completes.

Added a locking mechanism to the backup/restore functions. When a backup
or a restore function is executed it will check for a lock file and create it
if it does not exist. If it does exist, wait for 60 seconds while checking
to see if the lock has been removed. This will prevent a backup and restore
or multiple backups/restores from running simultaneously.

Also updated annotations for postgresql and mysql to match etcd.

Change-Id: I30090dbbf0e66cac133b6b2fa1d1b6690c127c78