Added STORAGE_POLICY env for both postgresql and etcd on-demand pods
REMOTE_BACKUP_STORAGE_POLICY is not being set in the on-demand pods, so if
the storage containers are created by the on-demand pods, they will not be
stored in the correct ceph pool.

The related backup cronjobs for each database type are already setup with
this value, this fix brings the on demand backup jobs in line with the
cronjobs.

Change-Id: I6fc83bdfb880bdf5841e12a72ce177ae2f364cdd