Fix etcd on-demand backup from failing due to no write access to hostpath mount
[0] introduced a breaking change to the etcd on demand database backup/restore
process. The on-demand pod uses a hostpath volume mount in which to store
the etcd backups. According to [1] hostpath mounts can only be written to
by the root user. The read only filesystem flag will remain in place and the
remaining filesystem will be read only, However, the volume mount will now be
writable again.

[0] https://review.opendev.org/#/c/749083/
[1] https://kubernetes.io/docs/concepts/storage/volumes/#hostpath

Change-Id: I7aa8c6bc32cddd889e3d16d68bc34e41db19b0af