Fix issue reading backup/restore secrets
This patchset uses jq to parse secrets instead of the grep/awk
combination that was being used before, which was problematic if the
formatting of secrets gets changed.

In order to do this for etcdctl-utility pod, I had to add "jq" to the
Dockerfile. Mysql and Postgresql utility images already had jq.

Change-Id: Ice7e7a44dbe9d6f8b4c7f02d2ed75c08ee47c89f