Add capability to add a test user/grants to test database
This patch will add the capability to add a test user and some
associated grants for the existing test database that gets added
when the database utility pods are started. DBUtils create_user,
query_user, and delete_user functions are provided so that a test
program can perform the needed functions to test backup/restore of
database users and corresponding grant information. Both MariaDB
and PostgreSQL test databases are updated for testing users/grants
in backup/restore.

In addition, capability to cleanup the test database backups is
added to reduce the test artifact footprint left by the utility
framework. The capability to delete a single backup will actually
be an operation that an operator/user can perform using dbutils
as well. This is added for mariadb/postgresql/etcd utility pods.

Change-Id: I683eb41c9a6ce2dc02018938a1d55a30bb87f3b3