Fix issue in dbutils where do_cleanup runs twice in a row
Dbutils for mysql currently experiences a small issues where it will
attempt to run the do_cleanup function twice in a row. This would occur
when running individual commands in non-interactive mode and using
the persistent flag -p for the on-demand pod.
Since dbutils requires a namespace to be entered for these commands
the code was failing to match and thus running an additional
do_cleanup when it was not needed.

This was also updated for postgresql so the script matches mysql.

Change-Id: I74c3bd948e7bf63caaa3249a486be818698e62b1