Correct small issue trying to get the ON_DEMAND pod name
There seems to be an unlikely case where it is possible the ON_DEMAND
pod name is not ready and we get a blank name, which in turn causes
additional failures.

Added a short 2s sleep to give the job additional time to create the
pod.

Added an additional check and error in case the pod name is still blank

Also refactored the code to use jq instead of grep | awk

Change-Id: Ibc33f1f0c8d37510b0112db8bb74c66cc595c890