Switch back to single node setup
This change reverts the multi-node setup to a
single node setup for the Porthole functional
tests. This is necessary to decrease the load on
the test environment and to simplify the
deployment process. In order to make it happen we
switch from rook-ceph to ceph chart from OSH,
which is more suitable for single node
deployments.

Change-Id: I085e6dd4f9d0a16d0abd183c0662de99b0e51823
Signed-off-by: Sergiy Markin <smarkin@mirantis.com>